+++ tests/test_erc.py
@@ -220,34 +220,56 @@ def test_all_pin_funcs_1():
 
 def test_assert_1():
 
-    res = Part(
-        tool=SKIDL,
-        name="res",
-        ref_prefix="R",
-        dest=TEMPLATE,
-        pins=[Pin(num=1, func=Pin.types.PASSIVE), Pin(num=2, func=Pin.types.PASSIVE)],
-    )
-    r1 = res(value="1K")
-    r2 = res(value="500")
+    erc_assert("len(Net.get('GND'))==1")  # Should fail.
 
-    cap = Part(
-        tool=SKIDL,
-        name="cap",
-        ref_prefix="C",
-        dest=TEMPLATE,
-        pins=[Pin(num=1, func=Pin.types.PASSIVE), Pin(num=2, func=Pin.types.NOCONNECT)],
-    )
-    c1 = cap()
-    c2 = cap(value="1uF")
+    r = Part("Device", "R", dest=TEMPLATE)
+    c = Part("Device", "C", dest=TEMPLATE)
+
+    r1 = r(value="1K")
+    r2 = r(value="500")
+    c1 = c()
+    c2 = c(value="1uF")
 
     gnd = Net("GND")  # Ground reference.
     vin = Net("VI")  # Input voltage to the divider.
     vout = Net("VO")  # Output voltage from the divider.
 
-    # add_erc_assertion(default_circuit, "len(gnd)==1")
-    # add_erc_assertion(default_circuit, compile("len(gnd)==0", "<stdin>", "eval"), globals(), locals())
-    add_erc_assertion(default_circuit, "len(gnd)==1")
+    erc_assert("len(gnd)==0", "Failed test!")  # Should not fail.
 
     ERC()
-    assert erc_logger.warning.count == 12
-    assert erc_logger.error.count == 0
+    assert erc_logger.warning.count == 14
+    assert erc_logger.error.count == 1
+
+
+def test_assert_2():
+
+    r = Part("Device", "R", dest=TEMPLATE)
+    c = Part("Device", "C", dest=TEMPLATE)
+
+    @subcircuit
+    def sub1(my_vin, my_gnd):
+        r1 = r()
+        c1 = c()
+        my_vin & r1 & c1 & my_gnd
+        erc_assert("len(my_vin) == len(my_gnd)")
+
+    @package
+    # @subcircuit
+    def sub2(my_vin, my_gnd):
+        sub1(my_vin, my_gnd)
+        sub1(my_vin, my_gnd)
+        erc_assert("len(my_vin) == 2")  # Will fail because of connection below.
+
+    vin, gnd = Net("VIN"), Net("GND")
+    sub = sub2()
+    vin += sub.my_vin
+    gnd += sub.my_gnd
+    # sub.my_vin += vin
+    # sub.my_gnd += gnd
+    # sub2(vin, gnd)
+    r1 = r()
+    vin & r1 & gnd  # Makes the assertion in sub2() fail.
+
+    ERC()
+
+    assert len(default_circuit.erc_assertion_list) == 3